+++ app/DefaultTags.jsx
@@ -1,0 +1,9 @@
+export default function DefaultTags() {
+  return (
+    <>
+      <meta content="width=device-width, initial-scale=1" name="viewport" />
+      <meta content="Generated by create next app" name="description" />
+      <link href="/favicon.ico" rel="icon" />
+    </>
+  );
+}
