+++ app/head.jsx
@@ -1,0 +1,10 @@
+import DefaultTags from './DefaultTags';
+
+export default function Head() {
+  return (
+    <>
+      <DefaultTags />
+      <title>SpotiFind</title>
+    </>
+  );
+}
