+++ app/layout.jsx
@@ -1,0 +1,18 @@
+import '@/styles/global.css';
+import SideBar from './SideBar';
+
+export default function RootLayout({ children }) {
+  return (
+    <html lang="en">
+      {/*
+        <head /> will contain the components returned by the nearest parent
+        head.js. Find out more at https://beta.nextjs.org/docs/api-reference/file-conventions/head
+      */}
+      <head />
+      <body>
+        <SideBar />
+        {children}
+      </body>
+    </html>
+  );
+}
