+++ app/page.jsx
@@ -1,0 +1,9 @@
+import styles from '@/styles/Home.module.css';
+
+export default function Home() {
+  return (
+    <main className={styles.main}>
+      <div className="text-red-500">I am Home</div>
+    </main>
+  );
+}
