+++ next.config.js
@@ -1,10 +1,13 @@
 /** @type {import('next').NextConfig} */
-const path = require('path');
+// const path = require('path');
 
 const nextConfig = {
   reactStrictMode: true,
-  sassOptions: {
-    includePaths: [path.join(__dirname, 'styles')]
+  // sassOptions: { // If using sass instead of postCSS
+  //   includePaths: [path.join(__dirname, 'styles')]
+  // },
+  experimental: {
+    appDir: true
   }
 };
