--- src/pages/_app.jsx
@@ -1,5 +1,0 @@
-import '@/styles/globals.scss';
-
-export default function App({ Component, pageProps }) {
-  return <Component {...pageProps} />;
-}
